requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.12.0,<4.0
djangorestframework-simplejwt>=5.0.0,<6.0
psycopg2-binary>=2.8.6,<3.0
django-filter>=2.4.0,<3.0
django-cors-headers>=3.7.0,<4.0
python-dotenv>=0.19.0,<1.0
drf-spectacular>=0.22.0,<1.0
django-debug-toolbar>=3.2.0,<4.0
redis
celery
